refact(Panels): create components for create panel
