+++ docker-compose.yml
@@ -32,7 +32,7 @@ services:
       dockerfile: Dockerfile
     volumes:
       - ./next/:/var/www
-    command: npm run dev
+    command: /bin/sh -c "npm install && npm run dev"
     labels:
       - "traefik.enable=true"
       - "traefik.http.routers.kudosboard.tls=true"
